Refactor `MusicPlayer` to inherit from new `VoiceClient` module
